fix: remove use of effect of state to avoid infinite loop & improve types for brn-column-manager
in the brn-toggle component we used an effect to sync the toggles state with the brn toggle group.
this effect caused an infinite loop when setting the state from the outside with ngModel.
used this chance to refactor the component and let the toggle group drive the updates.
also added improved types to the column manager that allows you toggle visibility of only passed in
columns.