fix: reflect disabled on host when input disabled is set to true or false
currently we simply pass on the disabled state to the cdkMenuElements but
do not set it on our brain host. This first showed as CSS issues, but could
also have other negative indication. So now we reflect the disabled input
in a binding to the host, which makes stiles from our hlm be applied correctly
and correctly marks the host as disabled if it is explicitly set.

Closes #60