feat(accordion): iterate on benni's accordion changes
couple things I added: make the hlm-accordion-icon a directive, too.
it is now an extension of the hlm-icon. I think this improves the
API surface a little. I wish we could also get rid of the brn-accordion-content component
and have hlm all the way, but I don't see an easy way to get around this. And I don't hate it,
it might just confuse some people. Other stuff is just minor cleanup, making what can be
readonly readonly, clean up some nested ifs, and reorganize the code in a way I think we can
use as a guideline.