fix(table): use some css dark magic to improve hover
we introduce a pseudo-element that we position behind the header row.
we then increase the children's z-index to be above the pseudo-element.
need to use 101 (and this might change in the future) as the cdk sticky.
sets it to 100. that way we still have a regular background and
no underlying table rows "shine through". when we hover we apply a semi
transparent muted background to get the same beautiful effect that shadcn
has.