fix(select): allow for null/undefined option
currently there is no way to have a null/undefined value as part of our options.
we should allow and display the correct option and therefore, we are only
resetting the selected options to an empty string and the value to null if
there isn't an actual option providing either of them.